task 7 was up date

+++ 0x13-more_singly_linked_lists/7-get_nodeint.c
@@ -7,7 +7,7 @@
  */
 listint_t *get_nodeint_at_index(listint_t *head, unsigned int index)
 {
-	size_t count = 0;
+	unsigned int count = 0;
 	listint_t *tmp = head;
 
 	if (head == NULL || index < 0)
